Refactoring. Re-use the same SubscriberObject type
