+++ src/modules/groups/groups.types.ts
@@ -1,5 +1,5 @@
 import { AxiosResponse } from "axios";
-import {Links, Meta} from "../../utils/types.js";
+import {SubscriberObject, Links, Meta} from "../../utils/types.js";
 
 export interface GroupsInterface {
     get:                (params: GetGroupsParams)                           => Promise<AxiosResponse<ListAllGroupsResponse>>;
@@ -71,34 +71,4 @@ interface GroupObject {
     bounced_count:      number;
     junk_count:         number;
     created_at:         string;
-}
-
-interface SubscriberObject {
-    id:               string;
-    email:            string;
-    status:           string;
-    source:           string;
-    sent:             number;
-    opens_count:      number;
-    clicks_count:     number;
-    open_rate:        number;
-    click_rate:       number;
-    ip_address:       string;
-    subscribed_at:    string;
-    unsubscribed_at:  string;
-    created_at:       string;
-    updated_at:       string;
-    fields: {
-        city:         string;
-        company:      string;
-        country:      string;
-        last_name:    string;
-        name:         string;
-        phone:        string;
-        state:        string;
-        z_i_p:        string;
-    };
-    groups:           Array<string>;
-    opted_in_at:      string;
-    optin_ip:         string;
 }
