+++ src/utils/types.ts
@@ -62,3 +62,34 @@ interface MetaLinks {
     label:  string;
     active: boolean;
 }
+
+export interface SubscriberObject {
+    id:               string;
+    email:            string;
+    status:           string;
+    source:           string;
+    sent:             number;
+    opens_count:      number;
+    clicks_count:     number;
+    open_rate:        number;
+    click_rate:       number;
+    ip_address:       string;
+    subscribed_at:    string;
+    unsubscribed_at:  string;
+    created_at:       string;
+    updated_at:       string;
+    fields: {
+        city:         string;
+        company:      string;
+        country:      string;
+        last_name:    string;
+        name:         string;
+        phone:        string;
+        state:        string;
+        z_i_p:        string;
+    };
+    groups:           Array<string>;
+    opted_in_at:      string;
+    optin_ip:         string;
+}
+
